extract detail::args as a reusable internal component
