+++ include/alloy.hpp
@@ -7,11 +7,12 @@
 
 #include "alloy/config.hpp"
 #include "alloy/append.hpp"
-#include "alloy/args.hpp"
 #include "alloy/at.hpp"
+#include "alloy/capture.hpp"
 #include "alloy/constant.hpp"
 #include "alloy/detail.hpp"
 #include "alloy/external.hpp"
+#include "alloy/forward.hpp"
 #include "alloy/join.hpp"
 #include "alloy/model.hpp"
 #include "alloy/prepend.hpp"
